Save what flashcards have been seen for up to 30 days
This was fun. To save what we see we convert the data into a bitstring
where a 0 in the 2 ^ position indicates that we saw the card at position
in the originals list. So 0 means we saw nothing, while a 1 would
indicate we saw only the first card in the list. The beauty of it of
course is that '010' means you saw the 2nd card, and '011' means you saw
the second and first card but it can be stored as the number 3. Which is
just the sum of the two 2 ^ position values. So it's pretty easy to
encode.

since we have over 900 cards, we need to represent 2 ^ 900 of course,
which is bigger than Javascript can handle on its own, hence the
BigInteger class we're using.

The BitInteger number is stored in a cookie for up to 30 days and is
decoded on load and saved each time we get a new card.